catalog: Move dialogs under EntityContextMenu
This removes back-and-forth state sharing between EntityLayout and
EntityContextMenu. This will also make it easier to make the header
for EntityLayout customizable because we don't need to somehow pass
the state through any possible customizations.
