+++ plugins/catalog/src/components/EntityLayout/EntityLayout.tsx
@@ -33,7 +33,6 @@ import {
   attachComponentData,
   IconComponent,
   useElementFilter,
-  useRouteRef,
   useRouteRefParams,
 } from '@backstage/core-plugin-api';
 import {
@@ -42,17 +41,13 @@ import {
   entityRouteRef,
   FavoriteEntity,
   getEntityRelations,
-  InspectEntityDialog,
-  UnregisterEntityDialog,
   useAsyncEntity,
 } from '@backstage/plugin-catalog-react';
 import Box from '@material-ui/core/Box';
 import { TabProps } from '@material-ui/core/Tab';
 import Alert from '@material-ui/lab/Alert';
-import React, { useEffect, useState } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
+import React from 'react';
 import { EntityContextMenu } from '../EntityContextMenu/EntityContextMenu';
-import { rootRouteRef, unregisterRedirectRouteRef } from '../../routes';
 import { catalogTranslationRef } from '../../translation';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
 
@@ -195,7 +190,6 @@ export const EntityLayout = (props: EntityLayoutProps) => {
   } = props;
   const { kind, namespace, name } = useRouteRefParams(entityRouteRef);
   const { entity, loading, error } = useAsyncEntity();
-  const location = useLocation();
   const routes = useElementFilter(
     children,
     elements =>
@@ -232,28 +226,7 @@ export const EntityLayout = (props: EntityLayoutProps) => {
     entity,
   );
 
-  const [confirmationDialogOpen, setConfirmationDialogOpen] = useState(false);
-  const [inspectionDialogOpen, setInspectionDialogOpen] = useState(false);
-  const navigate = useNavigate();
-  const catalogRoute = useRouteRef(rootRouteRef);
-  const unregisterRedirectRoute = useRouteRef(unregisterRedirectRouteRef);
   const { t } = useTranslationRef(catalogTranslationRef);
-
-  const cleanUpAfterRemoval = async () => {
-    setConfirmationDialogOpen(false);
-    setInspectionDialogOpen(false);
-    navigate(
-      unregisterRedirectRoute ? unregisterRedirectRoute() : catalogRoute(),
-    );
-  };
-
-  // Make sure to close the dialog if the user clicks links in it that navigate
-  // to another entity.
-  useEffect(() => {
-    setConfirmationDialogOpen(false);
-    setInspectionDialogOpen(false);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [location.pathname]);
 
   return (
     <Page themeId={entity?.spec?.type?.toString() ?? 'home'}>
@@ -268,8 +241,6 @@ export const EntityLayout = (props: EntityLayoutProps) => {
             <EntityContextMenu
               UNSTABLE_extraContextMenuItems={UNSTABLE_extraContextMenuItems}
               UNSTABLE_contextMenuOptions={UNSTABLE_contextMenuOptions}
-              onUnregisterEntity={() => setConfirmationDialogOpen(true)}
-              onInspectEntity={() => setInspectionDialogOpen(true)}
             />
           </>
         )}
@@ -300,18 +271,6 @@ export const EntityLayout = (props: EntityLayoutProps) => {
           )}
         </Content>
       )}
-
-      <UnregisterEntityDialog
-        open={confirmationDialogOpen}
-        entity={entity!}
-        onConfirm={cleanUpAfterRemoval}
-        onClose={() => setConfirmationDialogOpen(false)}
-      />
-      <InspectEntityDialog
-        open={inspectionDialogOpen}
-        entity={entity!}
-        onClose={() => setInspectionDialogOpen(false)}
-      />
     </Page>
   );
 };
